+++ platform/darwin/src/MGLTilePyramidOfflineRegion.h
@@ -44,6 +44,12 @@ MGL_EXPORT
  */
 @property (nonatomic, readonly) double maximumZoomLevel;
 
+/**
+ A specific list of tiles to download for this region.
+ Contains NSNumbers representing uint64_t tilekeys created by MGLTileKey
+*/
+@property (nonatomic, readonly, nullable) NSArray* tileList;
+
 - (instancetype)init NS_UNAVAILABLE;
 
 /**
@@ -74,6 +80,13 @@ MGL_EXPORT
     level.
  */
 - (instancetype)initWithStyleURL:(nullable NSURL *)styleURL bounds:(MGLCoordinateBounds)bounds fromZoomLevel:(double)minimumZoomLevel toZoomLevel:(double)maximumZoomLevel NS_DESIGNATED_INITIALIZER;
+
+/**
+  @param tileList An array of NSNumbers created by MGLTileKey representing the specific 
+    set of tiles to download for this region. If this is nil or empty a rectangular area
+    will be downloaded. 
+ */
+- (instancetype)initWithStyleURL:(NSURL *)styleURL bounds:(MGLCoordinateBounds)bounds fromZoomLevel:(double)minimumZoomLevel toZoomLevel:(double)maximumZoomLevel tileList:(nullable NSArray*)tileList;
 
 @end