+++ platform/ios/CHANGELOG.md
@@ -152,6 +152,7 @@ Mapbox welcomes participation and contributions from everyone. Please read [CONT
 * Fixed an issue that was causing the system location indicator to stay on in background after telemetry was disabled. ([#7833](https://github.com/mapbox/mapbox-gl-native/pull/7833))
 * Added support for predicates in rendered feature querying [8256](https://github.com/mapbox/mapbox-gl-native/pull/8246)
 * Added a nightly build of the dynamic framework. ([#8337](https://github.com/mapbox/mapbox-gl-native/pull/8337))
+* Added optional tileList parameter to MGLTilePyramidOfflineRegion to download non-rectangular offline packs, such as along a track.
 
 ## 3.4.2 - February 21, 2017